added entities

+++ servicetogo/src/main/java/com/greenfield/servicetogo/car/repository/CarServiceRequestRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.greenfield.servicetogo.car.entity.CarServiceRequestEntity;
 
-public interface CarServiceRequestRepository extends JpaRepository<CarServiceRequestEntity, Integer>  {
+public interface CarServiceRequestRepository extends JpaRepository<CarServiceRequestEntity, Long>  {
